==> safari_transparencia/notebooks/step1_crear_parquet_guardarS3.ipynb <==
import os, sys
sys.path.insert(0, os.path.abspath(".."))

import boto3
import pandas as pd
from src import etl as etl

## Importamos datos raw

Definimos el bucket de la S3 y el folder donde ese encuentran los archivos raw xls que importaremos con la información de 2012 a 2019.

bucket = 'inai-summerofdata'
folder = 'raw'

### Importamos 2012

f = f's3://{bucket}/{folder}/Solicitudes2012.xls'

df_2012 = pd.concat(pd.read_excel(f, sheet_name=None), ignore_index=True)

df_2012.head()

df_2012.info()

### Importamos 2013

f = f's3://{bucket}/{folder}/Solicitudes2013.xls'

df_2013 = pd.concat(pd.read_excel(f, sheet_name=None), ignore_index=True)

df_2013.head()

df_2013.info()

### Importamos 2014

f = f's3://{bucket}/{folder}/Solicitudes2014.xls'

df_2014 = pd.concat(pd.read_excel(f, sheet_name=None), ignore_index=True)

df_2014.head()

df_2014.info()

### Importamos 2015

f = f's3://{bucket}/{folder}/Solicitudes2015.xls'

df_2015 = pd.concat(pd.read_excel(f, sheet_name=None), ignore_index=True)

df_2015.head()

df_2015.info()

### Importamos 2016

f = f's3://{bucket}/{folder}/Solicitudes2016.xls'

df_2016 = pd.concat(pd.read_excel(f, sheet_name=None), ignore_index=True)

df_2016.head()

df_2016.info()

### Importamos 2017

f = f's3://{bucket}/{folder}/Solicitudes2017.xls'

df_2017 = pd.concat(pd.read_excel(f, sheet_name=None), ignore_index=True)

df_2017.head()

df_2017.info()

### Importamos 2018

f = f's3://{bucket}/{folder}/Solicitudes2018.xls'

df_2018 = pd.concat(pd.read_excel(f, sheet_name=None), ignore_index=True)

df_2018.head()

df_2018.info()

### Importamos 2019

fA = f's3://{bucket}/{folder}/Solicitudes2019-A.xls'
fB = f's3://{bucket}/{folder}/Solicitudes2019-B.xls'

df_2019A = pd.concat(pd.read_excel(fA, sheet_name=None), ignore_index=True)
df_2019B = pd.concat(pd.read_excel(fB, sheet_name=None), ignore_index=True)

df_2019 = df_2019A
df_2019 = df_2019.append(df_2019B, ignore_index=True)

df_2019.head()

df_2019.info()

Revisando la información de todos los años identificamos que algunos folios al importarlos pierden los ceros a la izquierda, consolidaremos todos los años en un solo dataframe y corregiremos los folios.

df = df_2012
df = df.append(df_2013, ignore_index=True)
df = df.append(df_2014, ignore_index=True)
df = df.append(df_2015, ignore_index=True)
df = df.append(df_2016, ignore_index=True)
df = df.append(df_2017, ignore_index=True)
df = df.append(df_2018, ignore_index=True)
df = df.append(df_2019, ignore_index=True)

df['FOLIO'] = df['FOLIO'].map(etl.Folio13)

df.head()

df.tail()

df.info()

## Descarga de archivos a la S3 y creación de las columnas tipo_archivo_respuesta y cantidad_archivos_respuesta

df['tipo_archivo_respuesta'], df['cantidad_archivos_respuesta'] = zip(*df['ARCHIVORESPUESTA'].map(etl.MimeType))

## Guardamos en formato parquet en S3

s3_url = f's3://{bucket}/parquet/inai.parquet'
df.astype({'FOLIO': str, 'CODIGOPOSTAL': str}).to_parquet(s3_url, engine='auto', compression='snappy')
