==> safari_transparencia/notebooks/step2_textract_pdf.ipynb <==
import os, sys
sys.path.insert(0, os.path.abspath(".."))

import os
import multiprocessing
import pandas as pd
import textract
import boto3
import botocore
from multiprocessing.dummy import Pool as ThreadPool
from src import etl as etl

## Importamos la información

Importamos los datos que se encuentran en formato parquet.

path = 's3://inai-summerofdata/parquet/legacy/inai_step1.parquet'
df = pd.read_parquet(path, engine='pyarrow')
pd.options.display.max_columns = None

Revisaremos las caracteristicas de nuestro dataframe.

df.head()

df.info()

## Extracción de texto haciendo paralelización de tareas.

Identificamos con cuantos cores cuenta nuestra máquina para aprovechar el máximo.

cores = multiprocessing.cpu_count()
pool = ThreadPool(cores)
cores

Creamos tuplas de los parámetros de la función que paralelizaremos.

tuples = [(folio, ext) for folio, ext in zip(df['FOLIO'], df['tipo_archivo_respuesta'])]

Ejecutamos la función para crear la columna texto_respuesta_adjunto que es la que contendra el texto extraído de los archivos pdf.

df['texto_respuesta_adjunto'] = pool.starmap(etl.TextractPDF, tuples)
pool.close()

df.head()

## Guardamos en formato parquet en S3

s3_url = f's3://inai-summerofdata/parquet/inai.parquet'
df.astype({'FOLIO': str, 'CODIGOPOSTAL': str}).to_parquet(s3_url, engine='auto', compression='snappy')
