==> safari_transparencia/notebooks/step4_fe_2.ipynb <==
import os, sys
sys.path.insert(0, os.path.abspath(".."))

import pandas as pd
from src import fe as fe

## Importamos la información

path = 's3://inai-summerofdata/mlpreproc/inai.parquet'
df = pd.read_parquet(path, engine='pyarrow')
pd.options.display.max_columns = None

Filtraremos solo por los casos que el tipo de solicitud sea "información pública" como lo definimos con SocialTIC.

df['tiposolicitud'].unique()

filtro = df['tiposolicitud'] == 'informacion publica'
df = df[filtro]

df.head()

df.info()

## Crearemos las variables calidadrespuesta y calidadrespuestareal

df['respuesta'].unique()

df = fe.CalidadRespuesta(df, 'respuesta', 'calidad_respuesta')

df = fe.CalidadRespuesta(df, 'respuestareal', 'calidad_respuesta_real')

df.head()

df.info()

Haremos un conteo de de los casos que tenemos en columna calidad_respuesta y calidad_respuesta_real

**calidad_respuesta**

grouped = df.groupby('calidad_respuesta').count()[['folio']]
grouped = grouped.sort_values('calidad_respuesta', ascending=False)
grouped.style.format("{:,}")

**calidad_respuesta_real**

grouped = df.groupby('calidad_respuesta_real').count()[['folio']]
grouped = grouped.sort_values('calidad_respuesta_real', ascending=False)
grouped.style.format("{:,}")

Podemos observar que aumentan al doble los casos "no respondida" para la columna calidad_respuesta_real.

## Guardamos en la S3 en la carpeta mlpreproc

s3_url = f's3://inai-summerofdata/mlpreproc/inai.parquet'
formatos = {'folio':str, 'fechasolicitud':'datetime64', 'dependencia':str,  'estatus':str,
           'medioentrada':str, 'tiposolicitud':str, 'descripcionsolicitud':str, 'otrosdatos':str,
           'archivoadjuntosolicitud':str, 'medioentrega':str, 'fechalimite':'datetime64',
           'respuesta':str, 'textorespuesta':str, 'archivorespuesta':str, 'fecharespuesta':'datetime64',
           'pais':str, 'estado':str, 'municipio':str, 'codigopostal':str, 'sector':str,
           'tipo_archivo_respuesta':str, 'cantidad_archivos_respuesta':'float64', 
            'texto_respuesta_adjunto':str, 'respuestareal':str, 'calidad_respuesta':str,
           'calidad_respuesta_real':str}
df.astype(formatos).to_parquet(s3_url, engine='auto', compression='snappy')
