==> safari_transparencia/notebooks/step5_model_LDA_0.ipynb <==
import os, sys
sys.path.insert(0, os.path.abspath(".."))

import pandas as pd

## Importamos la información

path = 's3://inai-summerofdata/mlpreproc/inai.parquet'
df = pd.read_parquet(path, engine='pyarrow')
pd.options.display.max_columns = None

df.head()

## Implementación LDA

Extracción de tópicos para textos libres de solicitudes.

Nos quedaremos con la columna descripcionsolictud de nuestro dataframe.

data = df[['descripcionsolicitud']]
data['index'] = data.index
docs = data

len(docs)

docs.head()

texto = docs[docs['index'] == 4310].values[0][0]
texto

### Pre-Procesamiento de la data

#### Tokenización

import spacy
nlp = spacy.load('es_core_news_sm')

texto = docs[docs['index'] == 4310].values[0][0]
texto

doc = nlp(texto)
doc

tokens = [t.orth_ for t in doc]
tokens

#### Limpieza del texto - tokenización

import spacy
nlp = spacy.load('es_core_news_sm')

texto = docs[docs['index'] == 4310].values[0][0]
texto

doc = nlp(texto)
doc

Quitamos puntos y stopwords.

lexical_tokens = [t.orth_ for t in doc if not t.is_punct | t.is_stop]
lexical_tokens

#### Normalizar texto

words = [t.lower() for t in lexical_tokens if len(t) > 3 and t.isalpha()]
words

### Creamos una función con que haga todo lo anterior - Tokenización (Normalizar texto)

def NormalizarTexto(texto):
    doc = nlp(texto)
    words = [t.orth_ for t in doc if not t.is_punct | t.is_stop]
    lexical_tokens = [t.lower() for t in words if len(t) > 3 and t.isalpha()]
    
    return lexical_tokens

word_list = NormalizarTexto(texto)
word_list

#### Lematización

import spacy
nlp = spacy.load('es_core_news_sm')

doc = nlp(texto)
doc

lemmas = [tok.lemma_.lower() for tok in doc]
lemmas

lemmas_no_pron = [tok.lemma_.lower() for tok in doc if tok.pos_ != 'PRON']
lemmas_no_pron

### Radicalizacion (stemming) -- Obtener raíces

import nltk
from nltk import SnowballStemmer

spanishstemmer = SnowballStemmer('spanish')

**tokenizamos**

tokens = NormalizarTexto(texto)
tokens

**steemming**

stems = [spanishstemmer.stem(token) for token in tokens]
stems

# Creamos una función para tokenizar, normalizar y radicalizar

import spacy
import nltk
from nltk import SnowballStemmer

nlp = spacy.load('es_core_news_sm')
spanishstemmer = SnowballStemmer('spanish')

def RadicalizarTexto(texto):
    
    doc = nlp(texto)
    words = [t.orth_ for t in doc if not t.is_punct | t.is_stop]
    lexical_tokens = [t.lower() for t in words if len(t) > 3 and t.isalpha()]
    
    stems = [spanishstemmer.stem(token) for token in lexical_tokens]
    
    return stems

**Ejemplo**

doc_sample = docs[docs['index'] == 4310].values[0][0]
doc_sample

word_list = RadicalizarTexto(doc_sample)
word_list

**Ahora lo haremos para todo el dataframe columna descripcionsolictud**

Paralelizaremos la tarea dado que es una tarea tardada.

import multiprocessing
from multiprocessing.dummy import Pool as ThreadPool

cores = multiprocessing.cpu_count()
pool = ThreadPool(cores)
cores

processed_docs = pool.map(RadicalizarTexto, docs['descripcionsolicitud'])
pool.close()

processed_docs[:10]

len(processed_docs)

Creamos una bag of words de nuestro dataframe

import gensim

dictionary = gensim.corpora.Dictionary(processed_docs)

count = 0
for k, v in dictionary.iteritems():
    print(k, v)
    count += 1
    if count > 10:
        break

Filtramos tokens que aparecen en:
+ Menos de 15 documentos o
+ mas que el 0.5 del total de documentos
+ despues de los dos pasos anteriores, manten solo los primeros 100,000 tokens mas frecuentes.

dictionary.filter_extremes(no_below=15, no_above=0.5, keep_n=100000)

**Gensim doc2bow**

Para cada documento creamos un diccionario reportando cuantas palabras y cuantas veces esas palabras aparecen.

bow_corpus = [dictionary.doc2bow(doc) for doc in processed_docs]

bow_corpus[4310]

### TF-IDF

Creamos un modelo tf-idf.

from gensim import corpora, models
from pprint import pprint

tfidf = models.TfidfModel(bow_corpus)
corpus_tfidf = tfidf[bow_corpus]

for doc in corpus_tfidf:
    pprint(doc)
    break

## Implementación LDA usando Bag of Words

lda_model = gensim.models.LdaMulticore(bow_corpus, num_topics=10,
                                      id2word=dictionary, passes=2,
                                      workers=cores)

Para cada topico exploraremos las palabras que contiene y sus pesos relativos.

for idx, topic in lda_model.print_topics(-1):
    print('Topic: {} \nWords: {}'.format(idx, topic))

## Implementación LDA usando TF-IDF

lda_model_tfidf = gensim.models.LdaMulticore(corpus_tfidf, 
                                            num_topics=10,
                                            id2word=dictionary,
                                            passes=2,
                                            workers=cores)

for idx, topic in lda_model_tfidf.print_topics(-1):
    print('Topic: {} Word: {}'.format(idx, topic))

## Evaluación de performance LDA Bag of Words -- Clasificaremos un ejemplo 

processed_docs[4310]

for index, score in sorted(lda_model[bow_corpus[4310]], 
                           key=lambda tup: -1*tup[1]):
    print("\nScore: {}\t \nTopic: {}".format(score,
                                            lda_model.print_topic(index, 10)))


for index, score in sorted(lda_model[bow_corpus[410]], 
                           key=lambda tup: -1*tup[1]):
    print("\nScore: {}\t \nTopic: {}".format(score,
                                            lda_model.print_topic(index, 10)))



## Evaluación de performance LDA TF-IDF -- Clasificaremos un ejemplo 

for index, score in sorted(lda_model_tfidf[bow_corpus[4310]],
                          key=lambda tup: -1*tup[1]):
    print("\nScore: {}\t \nTopic: {}.".format(score,
                                             lda_model_tfidf.print_topic(index, 10)))

for index, score in sorted(lda_model_tfidf[bow_corpus[410]],
                          key=lambda tup: -1*tup[1]):
    print("\nScore: {}\t \nTopic: {}.".format(score,
                                             lda_model_tfidf.print_topic(index, 10)))

# Guardamos como pickle

import pickle

with open('processed_docs_radicalizacion.pkl', 'wb') as f:
    pickle.dump(processed_docs, f)

with open('lda_model_bw_radicalizacion.pkl', 'wb') as f:
    pickle.dump(lda_model, f)

with open('lda_model_tfidf_radicalizacion.pkl', 'wb') as f:
    pickle.dump(lda_model_tfidf, f)

# Creamos una función para tokenizar, normalizar y lematizar.

import spacy
import nltk
from nltk import SnowballStemmer

nlp = spacy.load('es_core_news_sm')

def LematizarTexto(texto):
    
    doc = nlp(texto)
    words = [t.lemma_ for t in doc if not t.is_punct | t.is_stop]
    lexical_tokens = [t.lower() for t in words if len(t) > 3 and t.isalpha()]
    
    return lexical_tokens

texto

LematizarTexto(texto)

import multiprocessing
from multiprocessing.dummy import Pool as ThreadPool

cores = multiprocessing.cpu_count()
pool = ThreadPool(cores)
cores

processed_docs = pool.map(LematizarTexto, docs['descripcionsolicitud'])
pool.close()

processed_docs[:10]

import gensim

dictionary = gensim.corpora.Dictionary(processed_docs)

count = 0
for k, v in dictionary.iteritems():
    print(k, v)
    count += 1
    if count > 10:
        break

dictionary.filter_extremes(no_below=15, no_above=0.5, keep_n=100000)

bow_corpus = [dictionary.doc2bow(doc) for doc in processed_docs]

lda_model = gensim.models.LdaMulticore(bow_corpus, num_topics=10,
                                      id2word=dictionary, passes=2,
                                      workers=cores)

for idx, topic in lda_model.print_topics(-1):
    print('Topic: {} \nWords: {}'.format(idx, topic))

for index, score in sorted(lda_model[bow_corpus[4310]], 
                           key=lambda tup: -1*tup[1]):
    print("\nScore: {}\t \nTopic: {}".format(score,
                                            lda_model.print_topic(index, 10)))

from gensim import corpora, models
from pprint import pprint

tfidf = models.TfidfModel(bow_corpus)
corpus_tfidf = tfidf[bow_corpus]

lda_model_tfidf = gensim.models.LdaMulticore(corpus_tfidf, 
                                            num_topics=10,
                                            id2word=dictionary,
                                            passes=2,
                                            workers=cores)

for idx, topic in lda_model_tfidf.print_topics(-1):
    print('Topic: {} Word: {}'.format(idx, topic))

for index, score in sorted(lda_model_tfidf[bow_corpus[4310]],
                          key=lambda tup: -1*tup[1]):
    print("\nScore: {}\t \nTopic: {}.".format(score,
                                             lda_model_tfidf.print_topic(index, 10)))

import pickle

with open('processed_docs_lematizacion.pkl', 'wb') as f:
    pickle.dump(processed_docs, f)

with open('lda_model_bw_lematizacion.pkl', 'wb') as f:
    pickle.dump(lda_model, f)

with open('lda_model_tfidf_lematizacion.pkl', 'wb') as f:
    pickle.dump(lda_model_tfidf, f)

## Lematizando y quitando palabras [solicitar, informacion, mexico, titular, copiar, dependencia, municipio, ...]  (Segunda iteración)

import spacy
import nltk
from nltk import SnowballStemmer

nlp = spacy.load('es_core_news_sm')

def LematizarTexto(texto):
    
    doc = nlp(texto)
    words = [t.lemma_ for t in doc if not t.is_punct | t.is_stop]
    lexical_tokens = [t.lower() for t in words if len(t) > 3 and t.isalpha()]
    
    return lexical_tokens

banned = ['solicitar', 'solicito', 'solicitamos', 'solicite', 'informacion', 'mexico', 'titular', 
          'titulo', 'copiar', 'copie', 'copiamos', 'dependencia', 'municipio', 'tramitar', 'contratar', 
          'contrato', 'servicio', 'tramitar', 'tramite', 'proyectar', 'proyectamos', 'proyecte', 'proyeccion', 
          'nacional', 'administrativo', 'presupuestar', 'presupuesto', 'correo', 'pemex', 'programar', 'programa',
         'programacion', 'empresa']

f = lambda x: ' '.join([item for item in x.split() if item not in banned])
df['descripcionsolicitud'] = df['descripcionsolicitud'].apply(f)

data = df[['descripcionsolicitud']]
data['index'] = data.index
docs = data

texto = docs[docs['index'] == 4310].values[0][0]
texto

import multiprocessing
from multiprocessing.dummy import Pool as ThreadPool

cores = multiprocessing.cpu_count()
pool = ThreadPool(cores)
cores

processed_docs = pool.map(LematizarTexto, docs['descripcionsolicitud'])
pool.close()

import gensim

dictionary = gensim.corpora.Dictionary(processed_docs)

dictionary.filter_extremes(no_below=15, no_above=0.5, keep_n=100000)

bow_corpus = [dictionary.doc2bow(doc) for doc in processed_docs]

lda_model = gensim.models.LdaMulticore(bow_corpus, num_topics=10,
                                      id2word=dictionary, passes=2,
                                      workers=cores)

for idx, topic in lda_model.print_topics(-1):
    print('Topic: {} \nWords: {}'.format(idx, topic))

for index, score in sorted(lda_model[bow_corpus[4310]], 
                           key=lambda tup: -1*tup[1]):
    print("\nScore: {}\t \nTopic: {}".format(score,
                                            lda_model.print_topic(index, 10)))

from gensim import corpora, models
from pprint import pprint

tfidf = models.TfidfModel(bow_corpus)
corpus_tfidf = tfidf[bow_corpus]

lda_model_tfidf = gensim.models.LdaMulticore(corpus_tfidf, 
                                            num_topics=10,
                                            id2word=dictionary,
                                            passes=2,
                                            workers=cores)

for idx, topic in lda_model_tfidf.print_topics(-1):
    print('Topic: {} Word: {}'.format(idx, topic))

for index, score in sorted(lda_model_tfidf[bow_corpus[4310]],
                          key=lambda tup: -1*tup[1]):
    print("\nScore: {}\t \nTopic: {}.".format(score,
                                             lda_model_tfidf.print_topic(index, 10)))

import pickle

with open('processed_docs_lematizacion.pkl', 'wb') as f:
    pickle.dump(processed_docs, f)

with open('lda_model_bw_lematizacion.pkl', 'wb') as f:
    pickle.dump(lda_model, f)

with open('lda_model_tfidf_lematizacion.pkl', 'wb') as f:
    pickle.dump(lda_model_tfidf, f)
